% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{basic_scatter_plot}
\alias{basic_scatter_plot}
\title{basic_scatter_plot}
\usage{
basic_scatter_plot(df, x, y)
}
\arguments{
\item{df}{dataframe with the data to plot.}

\item{x}{variable to plot on the x axis.}

\item{y}{variable to plot on the y axis.}
}
\description{
\code{basic_scatter_plot} save a png file of a plot
}
