% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_annual_traits_R0_FT.R
\name{calculate_annual_traits_R0}
\alias{calculate_annual_traits_R0}
\title{model_param_list_create}
\usage{
calculate_annual_traits_R0(i, lthin, TS2, thinned, a_samps, PDR_samps,
  MDR_samps, e2a_samps, b_samps, c_samps, lf_samps, EFD_samps)
}
\arguments{
\item{i}{The data point}
}
\description{
\code{calculate_annual_traits_R0} calculates entomological traits for each
day of the year.
}
