% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df_to_list}
\alias{df_to_list}
\title{df_to_list}
\usage{
df_to_list(x, use_names)
}
\arguments{
\item{x}{the dataframe to convert.}

\item{use_names}{if TRUE it saves the df column names into each list component.}
}
\description{
\code{df_to_list} converts a dataframe into a list.
}
