% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loop_simplify}
\alias{loop_simplify}
\title{loop_simplify}
\usage{
loop_simplify(..., what)
}
\arguments{
\item{what}{the return value from \code{loop}.}
}
\description{
\code{loop_simplify} simplify the output of \code{loop}.
}
