% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_trait.R
\name{mean_trait}
\alias{mean_trait}
\title{mean_trait}
\usage{
mean_trait(my_data)
}
\arguments{
\item{my_data}{Dataframe of observations.}
}
\description{
\code{mean_trait} calculates the mean
of the trait values and standard deviation
}
