% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0_function.R
\name{myR0}
\alias{myR0}
\title{myR0}
\usage{
myR0(a, b, c, PDR, MDR, EFD, e2a, lf)
}
\arguments{
\item{a}{mosquitoes biting rate.}

\item{b}{proportion of infectious bites that infect susceptible humans.}

\item{c}{proportion of bites on infected humans that infect previously
uninfected mosquitoes.}

\item{PDR}{parasite development rate.}

\item{MDR}{mosquito immature development rate.}

\item{EFD}{number of eggs produced per female mosquito per day.}

\item{e2a}{mosquito egg-to-adult survival probability.}

\item{lf}{mosquito lifespan.}
}
\description{
\code{myR0} estimates virus reproduction number R0 based on a Ross-MacDonald model.
}
