% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thermal_responses.R
\name{plot_thermal_responses_mean_data}
\alias{plot_thermal_responses_mean_data}
\title{plot_thermal_responses_mean_data}
\usage{
plot_thermal_responses_mean_data(my_data_av, all_fits, out_pth, out_fl_nm)
}
\arguments{
\item{my_data_av}{Dataframe of average observations.}

\item{all_fits}{Dataframe of model fits.}

\item{out_pth}{The path where to save the plot}

\item{out_fl_nm}{The output file name}
}
\description{
\code{plot_thermal_responses_mean_data} makes a faceted plot of the
thermal responses with the mean of the observations.
}
