% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thermal_responses_uninform_priors.R
\name{plot_thermal_responses_uninform_priors}
\alias{plot_thermal_responses_uninform_priors}
\title{plot_thermal_responses_mean_data}
\usage{
plot_thermal_responses_uninform_priors(my_data_av, all_fits, out_pth)
}
\arguments{
\item{my_data_av}{Dataframe of average observations.}

\item{all_fits}{Dataframe of model fits.}

\item{out_pth}{The path where to save the plot}
}
\description{
\code{plot_thermal_responses_mean_data} makes a faceted plot of the
thermal responses with the mean of the observations.
}
