% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preds_vs_obs_scatter_plot.R
\name{preds_vs_obs_scatter_plot}
\alias{preds_vs_obs_scatter_plot}
\title{preds_vs_obs_scatter_plot}
\usage{
preds_vs_obs_scatter_plot(df, x, y, covar)
}
\arguments{
\item{df}{dataframe with the data to plot.}

\item{x}{variable to plot on the x axis.}

\item{y}{variable to plot on the y axis.}

\item{covar}{type of temperature (day or night) used to make predictions.}
}
\description{
\code{preds_vs_obs_scatter_plot} makes a simple scatter plot of predicted vs
oserved data points.
}
