% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{save_plot}
\alias{save_plot}
\title{save_plot}
\usage{
save_plot(plot_obj, out_pth, out_fl_nm, wdt, hgt)
}
\arguments{
\item{plot_obj}{The plot object}

\item{out_pth}{The path where to save the plot}

\item{out_fl_nm}{The output file name}

\item{wdt}{The plot width}

\item{hgt}{The plot height}
}
\description{
\code{save_plot} save a png file of a plot
}
