% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_out_csv}
\alias{write_out_csv}
\title{write_out_csv}
\usage{
write_out_csv(dat, my_path, file_name)
}
\arguments{
\item{dat}{the dataframe to save.}

\item{my_path}{the output directory.}

\item{file_name}{the output file name.}
}
\description{
\code{write_out_csv} saves a csv file.
}
