% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{write_out_rds}
\alias{write_out_rds}
\title{write_out_rds}
\usage{
write_out_rds(dat, my_path, file_name)
}
\description{
\code{write_out_rds} calculates entomological traits for each
day of the year.
}
